+++ src/main/java/io/github/zeroone3010/yahueapi/v2/domain/MotionResource.java
@@ -1,0 +1,31 @@
+package io.github.zeroone3010.yahueapi.v2.domain;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import io.github.zeroone3010.yahueapi.domain.JsonStringUtil;
+
+public class MotionResource extends Resource {
+  @JsonProperty("enabled")
+  private boolean enabled;
+
+  @JsonProperty("motion")
+  private Motion motion;
+
+  @JsonProperty("sensitivity")
+  private Sensitivity sensitivity;
+
+  public boolean isEnabled() {
+    return enabled;
+  }
+
+  public Motion getMotion() {
+    return motion;
+  }
+
+  public Sensitivity getSensitivity() {
+    return sensitivity;
+  }
+
+  public String toString() {
+    return JsonStringUtil.toJsonString(this);
+  }
+}